Fix undefined behaviour in the Buffer Manager after failure (#4246)
When multiple threads try to read the same page and run out of memory,
the page was being left as unlocked rather than evicted, which let other threads read from
the uninitialized data left over from the failed read.
In particular this was manifesting as segfaults in the Hash index if a
slot had no overflow slots, since the slot would be read as having a
nextOverflowSlot of 0 and it would attempt to read that nonexistant slot.

(cherry picked from commit 18c7e529ffbfbdbe498c05060b8e89b3684e2942)